backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
easyocr==1.7.1
opencv-python-headless==4.10.0.84
spacy==3.7.2
scispacy==0.5.3
torch==2.2.0
torchvision==0.17.0
numpy==1.26.2
Werkzeug==3.0.1
pandas==2.1.4
Pillow==10.1.0
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
nltk==3.8.1

# For scispaCy models, you may need:
# python -m spacy download en_core_web_sm
# pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_md-0.5.1.tar.gz